use a 'size_t' for representing the number of HTML items
Squashes 2 -Wconversion, -Wsign-conversion warnings.

+++ lib/common/htmltable.h
@@ -48,7 +48,7 @@ extern "C" {
     /* line of textspan_t's */
     typedef struct {
 	textspan_t *items;
-	short nitems;
+	size_t nitems;
 	char just;
 	double size;   /* width of span */
 	double lfsize; /* offset from previous baseline to current one */
